Create social share component
creates a component out of the right hand social media sharing buttons on blog posts
